backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
gunicorn>=23.0.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
httpx>=0.27.0
python-multipart>=0.0.9
slowapi>=0.1.9
email-validator>=2.1.0
structlog>=24.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
aiosqlite>=0.20.0
pgvector>=0.3.0
datasets>=3.0.0
tiktoken>=0.7.0
sse-starlette>=2.0.0
sentry-sdk[fastapi]>=2.0.0
